+++ DesktopApplicationTemplate.Tests/TestHelpers.cs
@@ -1,0 +1,14 @@
+using DesktopApplicationTemplate.Core.Converters;
+using DesktopApplicationTemplate.Models;
+using DesktopApplicationTemplate.UI.ViewModels;
+
+namespace DesktopApplicationTemplate.Tests;
+
+public static class TestHelpers
+{
+    public static ServiceListModel CreateService(ServiceType type, string name) => new ServiceListModel
+    {
+        ServiceType = type,
+        DisplayName = $"{ServiceTypeJsonConverter.ToLegacyString(type)} - {name}"
+    };
+}